+++ src/python/pants/backend/experimental/python/BUILD
@@ -2,8 +2,3 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_library()
-
-python_tests(
-  name = "tests",
-  timeout = 180,
-)
