+++ src/python/pants/backend/python/lint/black/subsystem_test.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 
 from textwrap import dedent
 
-from pants.backend.experimental.python.lockfile import PythonLockfileRequest
+from pants.backend.python.goals.lockfile import PythonLockfileRequest
 from pants.backend.python.lint.black import skip_field
 from pants.backend.python.lint.black.subsystem import Black, BlackLockfileSentinel
 from pants.backend.python.lint.black.subsystem import rules as subsystem_rules
