+++ src/python/pants/backend/python/subsystems/ipython_test.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 
 from textwrap import dedent
 
-from pants.backend.experimental.python.lockfile import PythonLockfileRequest
+from pants.backend.python.goals.lockfile import PythonLockfileRequest
 from pants.backend.python.subsystems.ipython import IPythonLockfileSentinel
 from pants.backend.python.subsystems.ipython import rules as subsystem_rules
 from pants.backend.python.target_types import PythonLibrary
