+++ src/python/pants/backend/python/subsystems/python_tool_base.py
@@ -6,9 +6,9 @@ from __future__ import annotations
 import importlib.resources
 from typing import ClassVar, Iterable, Sequence, cast
 
-from pants.backend.experimental.python.lockfile_metadata import calculate_invalidation_digest
 from pants.backend.python.target_types import ConsoleScript, EntryPoint, MainSpecification
 from pants.backend.python.util_rules.interpreter_constraints import InterpreterConstraints
+from pants.backend.python.util_rules.lockfile_metadata import calculate_invalidation_digest
 from pants.backend.python.util_rules.pex import Lockfile, LockfileContent, PexRequirements
 from pants.engine.fs import FileContent
 from pants.option.errors import OptionsError
